Major refactor to separate pages and recategorize components
