--- .eslintignore
@@ -1,49 +1,0 @@
-app/actions/actionTypes.js
-app/actions/authActions.js
-app/api/auth.js
-app/assets/tests/index.test.js
-app/components/Admin/Actions.js
-app/components/Admin/Admin.js
-app/components/Admin/ChangeRequest.js
-app/components/Admin/ChangeRequests.js
-app/components/Admin/ProposedService.js
-app/components/App.js
-app/components/Auth/RequireAuth.js
-app/components/Edit/EditAddress.js
-app/components/Edit/EditCollection.js
-app/components/Edit/EditNotes.js
-app/components/Edit/EditSchedule.js
-app/components/Find/CategoryItem.js
-app/components/Find/FindPage.js
-app/components/Footer.js
-app/components/Resource/CommunicationBoxes.js
-app/components/Resource/Hours.js
-app/components/Resource/ResourceMap.js
-app/components/Resource/Services.js
-app/components/Resource/DetailedHours.js
-app/components/Search/ResourcesMap.js
-app/components/Search/ResourcesRow.js
-app/components/User/CreateAccount.js
-app/components/User/Login.js
-app/components/User/Splash.js
-app/reducers/authReducer.js
-app/reducers/index.js
-app/reducers/resourceReducer.js
-app/store/configureStore.dev.js
-app/store/configureStore.js
-app/store/configureStore.prod.js
-app/utils/DataService.js
-app/utils/google.js
-app/actions/resourceActions.js
-app/components/Edit/EditAddress.js
-app/components/Edit/EditCollection.js
-app/components/Edit/EditNotes.js
-app/components/Edit/EditSchedule.js
-app/components/Edit/EditSections.js
-app/components/Edit/EditServices.js
-app/components/Edit/EditScheduleDay.js
-app/reducers/formConfig.js
-app/reducers/index.js
-app/routes.jsx
-app/store/*
-app/utils/index.js
