+++ app/App.jsx
@@ -1,9 +1,7 @@
 import React, { Component } from 'react';
-import Navigation from './ui/Navigation';
-import CategoryPage from './Find/FindPage';
-import ResourcesTable from './Search/ResourcesTable';
-import { round } from '../utils/index';
 import 'react-select/dist/react-select.css';
+import Navigation from 'components/ui/Navigation';
+import { round } from './utils/index';
 
 
 class App extends Component {
