+++ app/actions/actionTypes.js
@@ -2,6 +2,10 @@ export const ADMIN_LOGIN_SUCCESS = 'ADMIN_LOGIN_SUCCESS';
 export const ADMIN_LOGIN_ERROR = 'ADMIN_LOGIN_ERROR';
 export const ADMIN_LOGIN_REQUEST = 'ADMIN_LOGIN_REQUEST';
 
+export const ORGANIZATION_LOAD_REQUEST = 'ORGANIZATION_LOAD_REQUEST';
+export const ORGANIZATION_LOAD_SUCCESS = 'ORGANIZATION_LOAD_SUCCESS';
+export const ORGANIZATION_LOAD_FAIL = 'ORGANIZATION_LOAD_FAIL';
+
 export const SERVICE_LOAD_REQUEST = 'SERVICE_LOAD_REQUEST';
 export const SERVICE_LOAD_SUCCESS = 'SERVICE_LOAD_SUCCESS';
 export const SERVICE_LOAD_ERROR = 'SERVICE_LOAD_ERROR';
