--- app/actions/resourceActions.js
@@ -1,5 +1,0 @@
-export default {
-  addResource(resource) {
-
-  },
-};
