+++ app/components/edit/index.js
@@ -1,0 +1,6 @@
+export { default as EditAddress } from './EditAddress';
+export { default as EditServices } from './EditServices';
+export { default as EditNotes } from './EditNotes';
+export { default as EditSchedule } from './EditSchedule';
+export { default as EditPhones } from './EditPhones';
+export { default as EditSidebar } from './EditSidebar';
