+++ app/components/layout/index.js
@@ -1,3 +1,4 @@
 export { default as OrganizationCard } from './OrganizationCard';
 export { default as ServiceCard } from './ServiceCard';
 export { default as ListingTitleLink } from './ListingTitleLink';
+export { default as CategoryLink } from './CategoryLink';
