+++ app/components/listing/DetailedHours.js
