+++ app/components/listing/ResourceInfos.jsx
