+++ app/components/listing/TableOfOpeningTimes.jsx
@@ -9,7 +9,7 @@ class TableOfOpeningTimes extends React.Component {
     // TODO Show relativeOpeningTime for current day
     // TODO Show days without entries in the schedule as closed
     // TODO Order with current day at top
-    
+
     return (
       <table className="compact">
         <tbody>
