+++ app/components/ui/index.js
@@ -1,2 +1,3 @@
 export { default as Datatable } from './Datatable';
 export { default as Loader } from './Loader';
+export { default as MultiSelectDropdown } from './MultiSelectDropdown';
