+++ app/pages/LoginPage.jsx
@@ -1,10 +1,6 @@
 import React from 'react';
-import { Link } from 'react-router';
 import { connect } from 'react-redux';
-import { browserHistory } from 'react-router';
-import { images } from '../../assets';
-
-import authActions from '../../actions/authActions';
+import authActions from 'actions/authActions';
 
 class Login extends React.Component {
   constructor(props) {
