+++ app/reducers/resourceReducer.js
@@ -1,11 +1,14 @@
 import * as types from '../actions/actionTypes';
 
 const initialState = {
-  resource: null,
+  organizationList: [],
+  activeOrganization: null,
 };
 
 export default function resourceReducer(state = initialState, action) {
   switch (action.type) {
+    case types.ORGANIZATION_LOAD_SUCCESS:
+      return { ...state, activeOrganization: action.resource };
     default:
       return state;
   }
